# Conflicts:
#	src/main/java/com/a6raywa1cher/hackservspring/rest/exc/TokenIsNotEnabledException.java

+++ src/main/java/com/a6raywa1cher/hackservspring/rest/res/GetConfigResponse.java
@@ -7,7 +7,7 @@ public class GetConfigResponse {
 
     private String maxFileSize;
 
-    private Integer minEmailReq;
+    private Long minEmailReq;
 
-    private Integer maxEmailDuration;
+    private Long maxEmailDuration;
 }
